Special vibrational modes became OpenSystems
I have removed some problems with inheritance that made it impossible to make AnharmonicMode and HarmonicMode to become OpenSystems.
