+++ quantarhei/__init__.py
@@ -208,8 +208,8 @@ LIOUVILLE_PATHWAY_TYPES = PATHWAY_TYPES
 # Builders
 #
 from .builders.modes import Mode
-from .builders.submodes import AnharmonicMode
-from .builders.submodes import HarmonicMode
+from .builders.sysmodes import AnharmonicMode
+from .builders.sysmodes import HarmonicMode
 from .builders.vibsystem import VibrationalSystem
 from .builders.molecules import Molecule
 from .builders.molecule_test import TestMolecule
