+++ quantarhei/builders/__init__.py
@@ -1,8 +1,8 @@
 # -*- coding: utf-8 -*-
 
 from .modes import Mode
-from .submodes import AnharmonicMode
-from .submodes import HarmonicMode
+from .sysmodes import AnharmonicMode
+from .sysmodes import HarmonicMode
 from .molecules import Molecule
 from .aggregates import Aggregate
 from .pdb import PDBFile
